Rename module normalizing and create tests file.
