+++ packages/editor/src/client/blocks/richTextBlock/richTextBlock.tsx
@@ -10,7 +10,8 @@ import {RichTextBlockValue} from '../types'
 import {FormatButton, FormatIconButton, EditorSubMenuButton} from './toolbar/buttons'
 import {renderElement, renderLeaf} from './editor/render'
 import {BlockFormat, TextFormat} from './editor/formats'
-import {withRichText, withTable, withNormTables} from './editor/plugins'
+import {withRichText, withTable} from './editor/plugins'
+import {withNormTables} from './editor/normalizing'
 import {LinkFormatButton, RemoveLinkFormatButton} from './toolbar/linkButton'
 import {TableMenu} from './toolbar/tableMenu'
