+++ packages/editor/src/client/blocks/richTextBlock/toolbar/tableMenu.tsx
@@ -34,15 +34,15 @@ export function TableMenu() {
   const emptyCellsTable = (nrows: number, ncols: number): SlateElement[] => [
     {
       type: BlockFormat.Table,
-      children: Array.from({length: nrows}).map(() => ({
-        type: BlockFormat.TableRow,
-        children: Array.from({length: ncols}).map(() => ({
-          type: BlockFormat.TableCell,
-          borderColor: 'black',
-          // Wrap all content inside cell into paragraph block to enable break lines.
-          children: [emptyTextParagraph()]
-        }))
-      }))
+      // children: Array.from({length: nrows}).map(() => ({
+      //   type: BlockFormat.TableRow,
+      //   children: Array.from({length: ncols}).map(() => ({
+      //     type: BlockFormat.TableCell,
+      //     borderColor: 'black',
+      //     // Wrap all content inside cell into paragraph block to enable break lines.
+      children: [emptyTextParagraph()]
+      //   }))
+      // }))
     },
     // Append empty paragraph after table block for easy continuation.
     emptyTextParagraph()
